requirements.txt
aerich==0.5.8
aiomysql==0.0.22
aiosqlite==0.17.0
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.63.0
h11==0.12.0
idna==2.10
iso8601==0.1.16
pydantic==1.8.2
PyJWT==2.0.1
PyMySQL==0.9.3
pyparsing==3.0.6
pypika-tortoise==0.1.1
python-decouple==3.4
pytz==2021.3
redis==3.5.3
requests==2.25.1
starlette==0.13.6
tortoise-orm==0.17.8
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.13.4